添加 sglib与wtl两个开源库，chromium中加入ipc模块里面消息接收与处理子模块
- sglib a simple generic library for C.
- wtl Windows Template Library (WTL) is a C++ library for developing Windows applications and UI components. It extends ATL (Active Template Library) and provides a set of classes for controls, dialogs, frame windows, GDI objects, and more.
- chromium ipc 添加消息序列化，接收与应答宏，经过修改能够正常运行测试用例 ipc_sync_message_unittest

+++ chromium/base/time.cc
@@ -62,8 +62,8 @@ time_t Time::ToTimeT() const {
 
 // static
 Time Time::FromDoubleT(double dt) {
-  return (dt * static_cast<double>(kMicrosecondsPerSecond)) +
-      kTimeTToMicrosecondsOffset;
+    return static_cast<int64>((dt * static_cast<double>(kMicrosecondsPerSecond)) +
+        kTimeTToMicrosecondsOffset);
 }
 
 double Time::ToDoubleT() const {
